Relocate handler config setup for tests and utilized shared method
